Cleanup pass: fix 40+ bugs across core, operations, GUI, and CLI
Core engine:
- Forward Path values into subgraphs at run time (was silently skipped,
  leaving children with stale paths after save/reload)
- Blend hue along the shortest arc across the 0/360 seam in
  HSL/HSV/HWB/LCH/Oklch blend modes
- Compute FloatImage size arithmetic in usize (u32 overflow on huge dims)
- Emit ExposeInputChanged/ExposeOutputChanged (GUI handlers were dead)
- Iterative topological sort (recursion could overflow on deep graphs)

Operations:
- Fix panics: clamp with inverted/NaN bounds, from-hex and url-decode on
  non-ASCII input, random-integer at i32::MAX, NaN pixels in
  median/percentile, unbounded wired repeat/pad/format-number counts,
  blur on zero-width images, debug-build integer overflows
- Fix wrong output: top-hat/black-hat/morphological-gradient zeroed alpha
  on opaque RGBA; blit wrote grayscale fg into red channel only;
  off-origin crop edge-replicated instead of clipping; auto-levels no-op
  at low clip fractions; gradient last row never reached color B;
  normal-to-height scale input was a no-op; gabor kernel radius formula
  self-cancelled; lcm wrapped for large inputs
- Fix tiling seams for non-integer densities in gabor/phasor/craters and
  six grunge nodes; clamp fractal normalizers so persistence=0 is not NaN
- Reorder hydraulic erosion inputs to the simulation convention
  (seed/width/height first); document anisotropic noise 90-degree-only
  tiling honestly; dedupe smoothstep in histogram scan/select
- Make the material "uncreatable folder" test actually uncreatable on
  Windows (it was writing to the drive root)

GUI:
- Stop deep-copying PBR textures every frame in the 3D material preview
  (MaterialData now holds Arc<FloatImage>)
- Derive status-message and validation-error colors from the theme
- Remove dead double-click scaffold and a redundant per-connection lookup

CLI:
- Reject add-node/add-subgraph --id collisions instead of silently
  overwriting the existing node
- Exit non-zero from run/show-output when nodes error
- Warn when --save without --output drops extra outputs

~90 new regression tests.

+++ app/crates/mangler_core/src/float_image.rs
@@ -39,7 +39,7 @@ impl FloatImage {
     pub fn new(width: u32, height: u32, channels: u32) -> Self {
         assert!((1..=4).contains(&channels), "channels must be 1–4, got {}", channels);
         Self {
-            data: vec![0.0; (width * height * channels) as usize],
+            data: vec![0.0; width as usize * height as usize * channels as usize],
             width,
             height,
             channels,
@@ -55,7 +55,9 @@ impl FloatImage {
     pub fn from_pixel(width: u32, height: u32, channels: u32, pixel: &[f32]) -> Self {
         assert!((1..=4).contains(&channels), "channels must be 1–4, got {}", channels);
         assert_eq!(pixel.len(), channels as usize, "pixel length must match channels");
-        let total = (width * height) as usize;
+        // Compute the element count in usize so extreme dimensions can't
+        // overflow a u32 product (which would wrap and cause later OOB access).
+        let total = width as usize * height as usize;
         let mut data = Vec::with_capacity(total * channels as usize);
         for _ in 0..total {
             data.extend_from_slice(pixel);
@@ -66,7 +68,7 @@ impl FloatImage {
     /// Creates a FloatImage from raw data. Returns `None` if the data length
     /// does not match `width * height * channels`.
     pub fn from_raw(width: u32, height: u32, channels: u32, data: Vec<f32>) -> Option<Self> {
-        if data.len() != (width * height * channels) as usize {
+        if data.len() != (width as usize * height as usize * channels as usize) {
             return None;
         }
         Some(Self { data, width, height, channels })
@@ -202,7 +204,8 @@ impl FloatImage {
         // four specialized loops. Writing into a preallocated buffer through
         // zipped chunks keeps each iteration free of capacity checks so the
         // quantization can vectorize.
-        let mut out: Vec<u8> = vec![0u8; (self.width * self.height * 4) as usize];
+        // usize product avoids a u32 overflow on extreme dimensions.
+        let mut out: Vec<u8> = vec![0u8; self.width as usize * self.height as usize * 4];
         match self.channels {
             1 => {
                 for (dst, src) in out.chunks_exact_mut(4).zip(self.data.iter()) {
